extract types resolver as a independent pacakge
