+++ editor-packages/editor-services-estypes-resolver/package.json
@@ -1,0 +1,5 @@
+{
+  "name": "@code-editor/estypes-resolver",
+  "version": "0.0.0",
+  "private": false
+}
